backend/requirements.txt
amqp==5.1.1
APScheduler==3.10.0
asgiref==3.6.0
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
channels==3.0.5
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==39.0.1
daphne==3.0.2
defusedxml==0.7.1
distlib==0.3.6
Django==4.1.1
django-allauth==0.52.0
django-apscheduler==0.6.2
django-ckeditor==6.5.1
django-cors-headers==4.3.1
django-filter==22.1
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
filelock==3.9.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
inflection==0.5.1
kombu==5.2.4
oauthlib==3.2.2
packaging==24.0
Pillow==9.2.0
platformdirs==2.6.2
prompt-toolkit==3.0.38
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.6.0
pyOpenSSL==23.2.0
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
requests==2.28.2
requests-oauthlib==1.3.1
service-identity==23.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.1
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
virtualenv==20.17.1
wcwidth==0.2.6
zope.interface==6.0
